feat: add general email sender
