+++ email-service/package.json
@@ -13,11 +13,17 @@
   "description": "",
   "dependencies": {
     "axios": "^1.7.2",
-    "express": "^4.19.2"
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
+    "node-cron": "^3.0.3",
+    "nodemailer": "^6.9.14",
+    "winston": "^3.13.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.21",
     "@types/node": "^20.14.2",
+    "@types/node-cron": "^3.0.11",
+    "@types/nodemailer": "^6.4.15",
     "typescript": "^5.5.2"
   }
 }
