+++ email-service/src/services/email/email.service.interface.ts
@@ -2,9 +2,6 @@ export interface IEmailConfig {
     service: string
     login: string
     password: string
-    sender: string
-    subject: string
-    textTemplate: string
     port: number
 }
 
@@ -13,4 +10,8 @@ export interface IEmailDetails {
     to: string
     subject: string
     text: string
+}
+
+export interface IEmailSender{
+    sendEmail(emailDetails: IEmailDetails): Promise<void>
 }
