+++ email-service/src/services/scheduler/email.schedular.interface.ts
@@ -1,0 +1,6 @@
+export interface IRateEmailSchedularConfig {
+    textTemplate: string
+    sender: string
+    subject: string
+}
+
